requirements.txt
fschat==0.2.34
httpx==0.27.0
mosec==0.8.2
msgpack_python==0.5.6
nltk==3.8.1
numpy==1.24.1
openai==1.9.0
scikit_learn==1.3.2
tiktoken==0.5.2
torch==2.1.1+cu118
tqdm==4.66.1
transformers==4.36.0
